@@ -1,8 +1,9 @@
 {
-    "name": "goldpricetraking",
+    "name": "goldpricetraking-server",
     "version": "1.0.0",
     "description": "GoldPriceTracking with Node.js and IFTTT",
     "main": "index.js",
+    "private": true,
     "scripts": {
         "start": "tsc && node ./dist/src/index.js",
         "dev": "nodemon ./src/index.ts",
@@ -23,42 +24,44 @@
     "homepage": "https://github.com/oat9002/GoldPriceTraking#readme",
     "dependencies": {
         "async": "^3.2.2",
-        "axios": "^0.24.0",
+        "axios": "^0.25.0",
         "cheerio": "^1.0.0-rc.10",
         "cors": "^2.8.5",
         "dayjs": "^1.10.7",
-        "dotenv": "^10.0.0",
-        "eslint": "^8.3.0",
-        "express": "^4.17.1",
-        "firebase": "^9.5.0",
-        "firebase-admin": "^10.0.0",
+        "dotenv": "^14.2.0",
+        "eslint": "^8.7.0",
+        "express": "^4.17.2",
+        "firebase": "^9.6.3",
+        "firebase-admin": "^10.0.1",
         "jquery": "^3.6.0",
         "node-cron": "^3.0.0",
-        "qs": "^6.10.1",
+        "qs": "^6.10.3",
         "tracer": "^1.1.5",
-        "typescript": "^4.5.2"
+        "typescript": "^4.5.4"
     },
     "devDependencies": {
+        "@babel/core": "^7.16.7",
+        "@babel/eslint-parser": "^7.16.5",
         "@types/axios": "^0.14.0",
-        "@types/body-parser": "^1.19.1",
+        "@types/body-parser": "^1.19.2",
         "@types/cheerio": "^0.22.30",
         "@types/cors": "^2.8.12",
         "@types/dotenv": "^8.2.0",
         "@types/express": "^4.17.13",
         "@types/node-cron": "^3.0.0",
-        "@typescript-eslint/eslint-plugin": "^4.32.0",
-        "@typescript-eslint/parser": "^4.32.0",
-        "babel-eslint": "^10.1.0",
+        "@typescript-eslint/eslint-plugin": "^5.10.0",
+        "@typescript-eslint/parser": "^5.10.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-node": "^11.1.0",
         "eslint-plugin-prettier": "^4.0.0",
-        "nodemon": "^2.0.13",
-        "prettier": "^2.4.1",
-        "ts-node": "^10.2.1"
+        "nodemon": "^2.0.15",
+        "prettier": "^2.5.1",
+        "ts-node": "^10.4.0"
     },
     "eslintConfig": {
         "env": {
             "node": true
         }
-    }
+    },
+    "packageManager": "yarn@3.1.1"
 }
