@@ -1,32 +1,37 @@
 {
-    "name": "gold-price-tracking",
+    "name": "goldpricetraking-web",
     "version": "1.0.0",
+    "main": "src/index.js",
     "private": true,
     "devDependencies": {
-        "babel-eslint": "^10.1.0",
-        "eslint": "^7.32.0",
+        "@babel/core": "^7.16.7",
+        "@babel/eslint-parser": "^7.16.5",
+        "@babel/preset-react": "^7.16.7",
+        "babel-preset-react": "^6.24.1",
+        "eslint": "^8.7.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^4.0.0",
-        "eslint-plugin-react": "^7.26.1",
-        "prettier": "^2.4.1",
-        "react-scripts": "^4.0.0"
+        "eslint-plugin-react": "^7.28.0",
+        "eslint-plugin-react-hooks": "^4.3.0",
+        "prettier": "^2.5.1",
+        "react-scripts": "^5.0.0"
     },
     "dependencies": {
-        "@material-ui/core": "^4.11.0",
-        "@material-ui/icons": "^4.9.1",
+        "@material-ui/core": "^4.12.3",
+        "@material-ui/icons": "^4.11.2",
         "@material-ui/lab": "^4.0.0-alpha.56",
-        "axios": "^0.21.1",
-        "dayjs": "^1.10.3",
-        "firebase": "^8.2.4",
-        "prop-types": "^15.7.2",
-        "react": "^17.0.0",
-        "react-dom": "^17.0.0",
-        "react-redux": "^7.2.1",
-        "react-router-dom": "^5.2.0",
-        "recharts": "^2.0.9",
-        "redux": "^4.0.5",
-        "redux-thunk": "^2.3.0",
-        "styled-components": "^5.2.0"
+        "axios": "^0.24.0",
+        "dayjs": "^1.10.7",
+        "firebase": "^9.6.3",
+        "prop-types": "^15.8.1",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
+        "react-redux": "^7.2.6",
+        "react-router-dom": "^6.2.1",
+        "recharts": "^2.1.8",
+        "redux": "^4.1.2",
+        "redux-thunk": "^2.4.1",
+        "styled-components": "^5.3.3"
     },
     "scripts": {
         "start": "react-scripts start",
