@@ -1,4 +1,4 @@
-import { priceLastDay } from "mock/mock";
+import { priceLastDay } from "../mock/mock";
 import axios from "./Axios";
 
 export async function fetchGoldPrices(numOfDay) {
